[GUI] Make Schema a public type and add SchemaArrayInput
ToDo: implement SchemaArrayInput
